requirements.txt
# Core dependencies
fastapi>=0.104.1
uvicorn>=0.24.0
pydantic>=2.4.2
python-multipart>=0.0.6  # For file uploads
python-dotenv>=1.0.0

# Data processing and analysis
pandas>=2.1.2
numpy>=1.26.1
sqlite3-api>=0.1.0

# Visualization
matplotlib>=3.8.1
seaborn>=0.13.0

# LLM and AI
google-generativeai>=0.3.1

# Data loading
kagglehub>=0.2.0

# Optional but recommended
streamlit>=1.28.1  # For the frontend
requests>=2.31.0  # For API calls
typing-extensions>=4.8.0  # For type hints 

scikit-learn>=1.3.0  # For machine learning tasks
statsmodels>=0.14.0  # For statistical models
pillow>=10.0.0  # For image processing
scipy>=1.11.3  # For scientific computing